Move PQ client key share unit tests to their own file
